Create ReadMe.md
Capital Two Bank has issued a request for a proposal to be submitted on approaches to improve the model performance of its current logistic regression approach and simultaneously reduce modeling efforts. The outcome of this project recommends that Capital Two implement a k-means clustering algorithm with five clusters prior to logistic regression modeling and only utilize the labels generated from the clustering as the predictors. The recommended approach leads to an increase in average precision (+32.6%), recall (+7.7%), the consequent f1-score (+29.1%) and a decrease in misclassification costs (-48%); this approach not only increased model performance, but also removed the need for Capital Two to model all explanatory variables as features whilst logistic regression modeling.